refactor(core): move util into @flareapp/core, rename redactFullPath to redactUrlQuery
