+++ packages/core/src/util/convertToError.ts
