+++ packages/js/src/util/index.ts
@@ -1,8 +1,18 @@
-export * from './assert';
-export * from './convertToError';
-export * from './assertKey';
-export * from './extractCode';
-export * from './flatJsonStringify';
-export * from './glowsToEvents';
-export * from './now';
-export * from './redactUrl';
+export {
+    assert,
+    assertKey,
+    convertToError,
+    DEFAULT_URL_DENYLIST,
+    extractCode,
+    flatJsonStringify,
+    glowsToEvents,
+    now,
+    redactUrlQuery,
+    resolveDenylist,
+} from '@flareapp/core';
+
+// Re-export the original internal helpers that the rest of @flareapp/js relies on.
+// Internal callers continue to `import { ... } from '../util'` unchanged.
+
+/** @deprecated use redactUrlQuery instead - same behavior, honest name */
+export { redactUrlQuery as redactFullPath } from '@flareapp/core';
