+++ packages/js/tests/extractCode.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
 
-import { extractCode } from '../src/util/extractCode';
+import { extractCode } from '../src/util';
 
 describe('extractCode', () => {
     test('returns string code from error', () => {
